stat-service/requirements.txt
clickhouse-connect
grpcio
grpcio-tools
grpcio-reflection
